Rename Feature Extraction via HOG1.ipynb to Feature Extraction via HOG.ipynb
In this  task. HOG has been applied to compute the features and then apply two types of classifiers 
i.e Random Forest and SVM. During this task it was noticed that by changing the parameters of the model 
there was a significant change in the quantitative performance measures. Moreover, some of the 
examples were misclassified in the qualitative part. This was concluded by changing the parameters of 
HOG function from HOG (64,128) to HOG (128,128), setting iterations of SVM from 3000 to 1000 and 
Random Forest from 1000 to 2000. There was a significant change in the precision, recall and f1 score
along with weighted average of the model. On the basis of above mentioned, the model was misclassifying 
some images.